+++ golem-rib/src/type_inference/call_arguments_inference.rs
@@ -518,7 +518,7 @@ mod function_parameters_inference_tests {
         expr.infer_call_arguments_type(&function_type_registry)
             .unwrap();
 
-        let let_binding = Expr::let_binding("x", Expr::untyped_number(1f64));
+        let let_binding = Expr::let_binding("x", Expr::untyped_number(BigDecimal::from(1)));
 
         let call_expr = Expr::Call(
             CallType::Function(DynamicParsedFunctionName {
